% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggite.r
\name{aggite}
\alias{aggite}
\title{Aggregate Unit-Time Average Treatment Effects}
\usage{
aggite(
  MP,
  type = "group",
  balance_e = NULL,
  min_e = -Inf,
  max_e = Inf,
  min_agg = 2,
  na.rm = FALSE,
  bstrap = NULL,
  biters = NULL,
  cband = NULL,
  alp = NULL,
  clustervars = NULL
)
}
\arguments{
\item{MP}{an MP_i object (i.e., the results of the \code{\link[=att_it]{att_it()}} method)}

\item{type}{Which type of aggregated treatment effect parameter to compute.
One option is "simple" (this just computes a weighted average of all
unit-time average treatment effects).
Other options are "dynamic" (this computes average effects across
different lengths of exposure to the treatment and is similar to an
"event study"; here the overall effect averages the effect of the
treatment across all positive lengths of exposure); "group" (this
is the default option and
computes average treatment effects across different groups; here
the overall effect averages the effect across different groups); and
"calendar" (this computes average treatment effects across different
time periods; here the overall effect averages the effect across each
time period).}

\item{balance_e}{If set (and if one computes dynamic effects), it balances
the sample with respect to event time.  For example, if \code{balance.e=2},
\code{aggite} will drop groups that are not exposed to treatment for
at least three periods. (the initial period when \code{e=0} as well as the
next two periods when \code{e=1} and the \code{e=2}).  This ensures that
the composition of groups does not change when event time changes.}

\item{min_e}{For event studies, this is the smallest event time to compute
dynamic effects for.  By default, \code{min_e = -Inf} so that effects at
all lengths of exposure are computed.}

\item{max_e}{For event studies, this is the largest event time to compute
dynamic effects for.  By default, \code{max_e = Inf} so that effects at
all lengths of exposure are computed.}

\item{min_agg}{the minimum number of unit-time effects required for aggregation. Default is 2.}

\item{na.rm}{Logical value if we are to remove missing Values from analyses. Defaults is FALSE.}

\item{bstrap}{This is always TRUE. Turning off makes no difference in the second-step.}

\item{biters}{The number of bootstrap iterations to use.  The default is the value set in the MP_i object,
and this is only applicable if \code{bstrap=TRUE}.}

\item{cband}{Does not apply. All confidence intervals are bootstrapped confidence intervals in the second-step.}

\item{alp}{the significance level, default is value set in the MP object.}

\item{clustervars}{ignored in this function.}
}
\value{
An \code{\link{AGGITEobj}} object that holds the results from the
aggregation
}
\description{
A function to take unit-time average treatment effects
and aggregate them into a smaller number of parameters.  There are
several possible aggregations including "simple", "dynamic", "unit,
"group","dosage","calendar, and a custom aggregation name."
}
\examples{
# first run the att_it() function
simdata = sim_data(posttreat=6)
attobject = att_it(yname = "y", tname = "time", gname = "treatg", idname ="unit", data = simdata)

# aggregate all post-treatment effects of each unit
agtobject = aggite(attobject,type="unit")


}
