% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attit_table.R
\name{attit_table}
\alias{attit_table}
\title{Issues tabulated att_it() results}
\usage{
attit_table(MP)
}
\arguments{
\item{MP}{an \code{MP_i} object that contains the results from an \code{att_it()} call}
}
\value{
a dataframe with the id, group, time, att estimates, standard errors and confidence intervals.
}
\description{
Issues tabulated att_it() results
}
\examples{
simdata = sim_data()
attobject = att_it(yname = "y", tname = "time", gname = "treatg", idname ="unit", data = simdata)
attdf = attit_table(attobject)


}
