% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.aggite.R
\name{compute.aggite}
\alias{compute.aggite}
\title{Compute Aggregated Treatment Effect Parameters}
\usage{
compute.aggite(
  MP,
  type = "group",
  balance_e = NULL,
  min_e = -Inf,
  max_e = Inf,
  min_agg = 2,
  na.rm = FALSE,
  bstrap = NULL,
  biters = NULL,
  maxbackp = 5,
  ignorebackp = FALSE,
  cband = NULL,
  alp = NULL,
  clustervars = NULL,
  call = NULL
)
}
\arguments{
\item{MP}{an MP_i object (i.e., the results of the \code{\link[=att_it]{att_it()}} method)}

\item{type}{Which type of aggregated treatment effect parameter to compute.
One option is "simple" (this just computes a weighted average of all
unit-time average treatment effects).
Other options are "dynamic" (this computes average effects across
different lengths of exposure to the treatment and is similar to an
"event study"; here the overall effect averages the effect of the
treatment across all positive lengths of exposure); "group" (this
is the default option and
computes average treatment effects across different groups; here
the overall effect averages the effect across different groups); and
"calendar" (this computes average treatment effects across different
time periods; here the overall effect averages the effect across each
time period).}

\item{balance_e}{If set (and if one computes dynamic effects), it balances
the sample with respect to event time.  For example, if \code{balance.e=2},
\code{aggite} will drop groups that are not exposed to treatment for
at least three periods. (the initial period when \code{e=0} as well as the
next two periods when \code{e=1} and the \code{e=2}).  This ensures that
the composition of groups does not change when event time changes.}

\item{min_e}{For event studies, this is the smallest event time to compute
dynamic effects for.  By default, \code{min_e = -Inf} so that effects at
all lengths of exposure are computed.}

\item{max_e}{For event studies, this is the largest event time to compute
dynamic effects for.  By default, \code{max_e = Inf} so that effects at
all lengths of exposure are computed.}

\item{min_agg}{the minimum number of unit-time effects required for aggregation. Default is 2.}

\item{na.rm}{Logical value if we are to remove missing Values from analyses. Defaults is FALSE.}

\item{bstrap}{This is always TRUE. Turning off makes no difference in the second-step.}

\item{biters}{The number of bootstrap iterations to use.  The default is the value set in the MP_i object.}

\item{maxbackp}{The number of pre-period observations sampled for baseline variance. Default is 5.}

\item{ignorebackp}{Logical for whether the pre-period should be sampled for baseline variance. Default is FALSE.}

\item{cband}{Does not apply. All confidence intervals are bootstrapped confidence intervals in the second-step.}

\item{alp}{the significance level, default is value set in the MP_i object.}

\item{clustervars}{ignored in this function.}

\item{call}{The function call to aggite}
}
\value{
\code{\link{AGGITEobj}} object
}
\description{
Computes aggregated average treatment effects
}
\examples{
# This is a helper function for [aggite()]. See that function for examples.


}
\keyword{internal}
