% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.att_it.R
\name{compute.att_it}
\alias{compute.att_it}
\title{Compute Group-Time Average Treatment Effects}
\usage{
compute.att_it(dp)
}
\arguments{
\item{dp}{a DIDparams_i object}
}
\value{
a list object with the calculated att_it() and corresponding influence functions
}
\description{
\code{compute.att_it} does the main work for computing
multiperiod group-time average treatment effects while balancing for the effects at each time period
}
