% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIDparams_i.R
\name{DIDparams_i}
\alias{DIDparams_i}
\title{Title DIDparams_i}
\usage{
DIDparams_i(
  yname,
  tname,
  idname,
  gname,
  cohort = NULL,
  xformla = NULL,
  customnames = NULL,
  data,
  control_group,
  anticipation = 0,
  weightsname = NULL,
  weightfs = FALSE,
  alp = 0.05,
  bstrap = TRUE,
  biters = 1000,
  clustervars = NULL,
  cband = TRUE,
  print_details = TRUE,
  pl = FALSE,
  cores = 1,
  est_method = "dr",
  overlap = "trim",
  base_period = "varying",
  panel = TRUE,
  fixedbase = NULL,
  nobase = FALSE,
  n = NULL,
  nG = NULL,
  nT = NULL,
  tlist = NULL,
  glist = NULL,
  idlist = NULL,
  call = NULL
)
}
\arguments{
\item{yname}{the name of the outcome variable in the \code{data}}

\item{tname}{the name of the column containing the time periods in \code{data}. Must be numeric}

\item{idname}{the individual (cross-sectional unit) id name in \code{data}. Cannot be missing and must be numeric.}

\item{gname}{the name of the variable in \code{data} that
contains the first period when a particular observation is treated.
This should be a positive number for all observations in treated groups.
It defines which "treatment-group" a unit belongs to. A zero (0) value is considered to be a unit that is never treated.
A unit treated after the last period in \code{tname} is also considered never treated.}

\item{cohort}{the name of blocking variable in \code{data}, specified if comparisons should only be within pre-treatment cohort.
If blocking is required across the entire trajectory, this must be imposed on the input to \code{data} prior to running this function.}

\item{xformla}{a formula for the covariates to include in the
model.  It should be of the form \code{~ X1 + X2}. An intercept is automatically included. Default
is NULL which is equivalent to \code{xformla=~1}.}

\item{customnames}{the names of additional aggregation variables in \code{data}.}

\item{data}{the name of the data.frame that contains the data.}

\item{control_group}{which units to use the control group.
The default is "nevertreated" which sets the control group
to be the group of units that never participate in the
treatment.  This group does not change across groups or
time periods.  The other option is to set
\code{group="notyettreated"}.  In this case, the control group
is set to the group of units that have not yet participated
in the treatment in that time period.  This includes all
never treated units, but it includes additional units that
eventually participate in the treatment, but have not
participated yet.}

\item{anticipation}{the number of time periods before participating
in the treatment where units can anticipate participating in the
treatment and therefore it can affect their untreated potential outcomes.}

\item{weightsname}{the name of the column containing the sampling weights.
If not set, all observations have same weight.}

\item{weightfs}{boolean for whether the att's should be weighted. Default
is FALSE so that only the aggite step is weighted unless specified.}

\item{alp}{the significance level, default is 0.05}

\item{bstrap}{a deprecated option. Only analytical standard errors are reported.}

\item{biters}{a deprecated option, also see \code{bstrap}.}

\item{clustervars}{a deprecated option, also see \code{bstrap}.}

\item{cband}{a deprecated option. Does not produce uniform confidence bands.}

\item{print_details}{whether or not to show details/progress of computations.
Default is \code{FALSE}.}

\item{pl}{whether or not to use parallel processing}

\item{cores}{the number of cores to use for parallel processing}

\item{est_method}{the method to compute group-time average treatment effects.  The default is "dr" which uses the doubly robust
approach in the \code{DRDID} package.  Other built-in methods
include "ipw" for inverse probability weighting (Hajek type) and "reg" for
first step regression estimators.}

\item{overlap}{the treatment of units that violate overlap conditions when the \code{est_method} is "dr" or "ipw".
The default, "trim", is to drop the unit but report the calculated ATT for further analysis. Overlap is violated if the maximum pscore exceeds 0.999
The other option, "retain" retains these units for inference.}

\item{base_period}{whether to use a "varying" base period or a
"universal" base period for placebo tests. A varying base period calculates a pseudo-ATT for every two consecutive pre-treatment periods.
A universal base period fixes the base period to always be (g-anticipation-1).  Either choice results in the same
post-treatment estimates of ATT(i,t)'s.}

\item{panel}{if TRUE, the data is coerced to a balanced panel on the. Default is TRUE.}

\item{fixedbase}{if the user needs to fix a base period to be a particular period, use this option. Must be numeric. Default is NULL.}

\item{nobase}{if TRUE, the baseline vectors are set to zero. Estimates are not differences-in-differences. Default is FALSE.}

\item{n}{The number of individual and time observations}

\item{nG}{The number of units and not just treatment cohorts}

\item{nT}{The number of time periods}

\item{tlist}{A vector containing all time periods}

\item{glist}{A vector containing each treatment-time group and allows repetitions.}

\item{idlist}{A vector containing each id that corresponds to the glist vector.}

\item{call}{Function call to att_it}
}
\value{
a DIDparams_i object
}
\description{
Title DIDparams_i
}
\examples{
# Helper function for [pre_process_did_i()]. See documentation of that function for an example.


}
