% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGGITEobj.r
\name{print.AGGITEobj}
\alias{print.AGGITEobj}
\title{print.AGGITEobj}
\usage{
\method{print}{AGGITEobj}(x, ...)
}
\arguments{
\item{x}{a \code{AGGITEobj} object}

\item{...}{extra arguments}
}
\description{
prints value of a \code{AGGITEobj} object
}
