% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_i.r
\name{print.MP_i}
\alias{print.MP_i}
\title{print.MP_i}
\usage{
\method{print}{MP_i}(x, ...)
}
\arguments{
\item{x}{a \code{MP_i} object}

\item{...}{extra arguments}
}
\description{
prints value of a \code{MP_i} object
}
