% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_attit.r
\name{process_attit}
\alias{process_attit}
\title{Process Results from \code{\link[=compute.att_it]{compute.att_it()}}}
\usage{
process_attit(attgt.list)
}
\arguments{
\item{attgt.list}{list of results from \code{\link[=compute.att_it]{compute.att_it()}}}
}
\value{
list with elements:
\item{id}{which id a set of results belongs to}
\item{group}{which group the id belongs to}
\item{tt}{which time period a set of results belongs to}
\item{att}{the unit time average treatment effect (NA if there were propensity score problems)}
\item{se}{the standard error of the unit time average treatment effect (NA if there were propensity score problems)}
\item{lci}{the lower confidence interval of the unit time average treatment effect (NA if there were propensity score problems)}
\item{uci}{the upper confidence interval of the unit time average treatment effect (NA if there were propensity score problems)}
\item{ipwqual}{the maximum propensity score}
\item{baseline}{the pre-treatement outcome value}
\item{attcalc}{the unit time average treatment effect (always complete unless NaN)}
}
\description{
Process Results from \code{\link[=compute.att_it]{compute.att_it()}}
}
\examples{
# A helper function of [att_it()], but can be useful for debugging.


}
