% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_ipw_did_panel_archive.R
\name{std_ipw_did_panel_archive}
\alias{std_ipw_did_panel_archive}
\title{An archived version of the DRDID's std_ipw_did_panel without the fastglm function}
\usage{
std_ipw_did_panel_archive(
  y1,
  y0,
  D,
  covariates,
  i.weights = NULL,
  boot = FALSE,
  boot.type = "weighted",
  nboot = NULL,
  inffunc = FALSE
)
}
\arguments{
\item{y1}{The post-treatment vector}

\item{y0}{The pre-treatment vector}

\item{D}{The binary treatment vector}

\item{covariates}{The matrix or vector of covariates to train weights}

\item{i.weights}{Sample weights. Default is NULL.}

\item{boot}{A logical as to whether to bootstrap. Default is FALSE.}

\item{boot.type}{Type of bootstrap. Default is a weighted bootstrap.}

\item{nboot}{Number of bootstraps}

\item{inffunc}{A logical as to whether to report the influence function or not.}
}
\value{
Returns an object with calculated ATT, standard error, confidence intervals, and influence functions.
}
\description{
An archived version of the DRDID's std_ipw_did_panel without the fastglm function
}
\examples{
std_ipw_did_panel_archive(rep(1,20),rep(0,20),c(1,rep(0,19)),c(70,rep(50,9),rep(80,10)))

}
