% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector with, at most, one element.}

\item{pattern}{A regex pattern}

\item{n}{number of splits to make}
}
\value{
A character vector.
}
\description{
Split a string
}
