% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGGITEobj.r
\name{summary.AGGITEobj}
\alias{summary.AGGITEobj}
\title{Summary Aggregate Treatment Effect Parameter Objects}
\usage{
\method{summary}{AGGITEobj}(object, ...)
}
\arguments{
\item{object}{an \code{AGGITEobj} object}

\item{...}{other arguments}
}
\description{
A function to summarize aggregated treatment effect parameters.
}
