% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_i.r
\name{summary.MP_i}
\alias{summary.MP_i}
\title{summarizes an MP_i object}
\usage{
\method{summary}{MP_i}(object, ...)
}
\arguments{
\item{object}{an \code{MP_i} object}

\item{...}{extra arguments}
}
\description{
prints a summary of a \code{MP_i} object
}
