% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggite_table.R
\name{aggite_table}
\alias{aggite_table}
\title{Issues tabulated aggite() results}
\usage{
aggite_table(AGGITE)
}
\arguments{
\item{AGGITE}{an object of class AGGITEobj that collects the results from a \code{aggite()} call}
}
\value{
a data-frame with the aggregation type \code{egt},
the corresponding att calculation \code{att.egt}, standard errors \code{se.egt}, and confidence intervals.
}
\description{
Issues tabulated aggite() results
}
