% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_i.r
\name{MP_i}
\alias{MP_i}
\title{MP_i}
\usage{
MP_i(
  id,
  group,
  t,
  att,
  V_analytical,
  se,
  c,
  lci,
  uci,
  inffunc,
  n = NULL,
  aggite = NULL,
  alp = 0.05,
  ipwqual = NULL,
  attcalc = NULL,
  baseline = NULL,
  count = NULL,
  DIDparams = NULL
)
}
\arguments{
\item{id}{which unit the unit-time treatment effects are for}

\item{group}{which group (defined by period first treated) an unit-time average treatment effect is for}

\item{t}{which time period a group-time average treatment effect is for}

\item{att}{the estimate for unit \code{id} and time
period \code{t}}

\item{V_analytical}{Analytical estimator for the asymptotic variance-covariance matrix for unit-time average treatment effects}

\item{se}{standard errors for unit-time average treatment effects. If bootstrap is set to TRUE, this provides bootstrap-based se.}

\item{c}{always NULL}

\item{lci}{lower confidence interval for att}

\item{uci}{upper confidence interval for att}

\item{inffunc}{the influence function for estimating group-time average treatment effects}

\item{n}{the number of unique cross-sectional units (unique values of idname)}

\item{aggite}{an aggregate treatment effects object}

\item{alp}{the significance level, default is 0.05}

\item{ipwqual}{the maximum propensity score}

\item{attcalc}{similar to att but does not remove estimates when there are propensity score problems}

\item{baseline}{the baseline of the unit}

\item{count}{takes a count of the number of units in the component estimate}

\item{DIDparams}{a \code{\link{DIDparams_i}} object.  A way to optionally return the parameters
of the call to \code{\link[=att_it]{att_it()}}.}
}
\value{
MP object
}
\description{
Multi-period objects that hold results for unit-time average treatment effects
}
\examples{
# Helper function for [att_it()]. See documentation of that function for an example.


}
